Remove external links styling
That experiment for short long-term evaluation whether or not external
link style is worth it or not is over. And I think it's too much
considering that the target audience is for personal blogs which will
not mind such things. Plus the amount of things to make it work properly
is just too much (i.e., the font has to have the glyph).